put the material UI theme objects into own files
