+++ src/components/App/App.js
@@ -1,42 +1,20 @@
 import React from 'react';
-import Dashboard from'../Dashboard/Dashboard.js';
+import Dashboard from'../DashBoard/Dashboard.js';
 import {BrowserRouter as Router,Switch,Route} from "react-router-dom";
-import Products from '../componentsUtils/Products.js';
-import Test from '../Test.js';
+import Products from '../../componentsUtils/Products/Products.js';
+import Test from '../Test/Test.js';
 import Prof from '../Prof/Prof.js';
-import Doc from '../../componentsUtils/Doc.js';
+import Doc from '../../componentsUtils/Docs/Doc.js';
 
 function App() {
     return (
         <Router>
             <Switch>
-                <Route path="/home">
-                    <div className="dash">
-                        <Dashboard />
-                        {/* <Front /> */}
-                    </div>
-                </Route>
-                <Route exact path="/prof">
-                    <div className="prof">
-                        <Prof />
-                    </div>
-                </Route>
-                <Route path="/products">
-                    <div className="products">
-                        <Products />
-                    </div>
-                </Route>
-                <Route path="/test">
-                    <div className="test">
-                        <Test />
-                    </div>
-                </Route>
-                <Route path="/doc">
-                    <div className="test">
-                        <Doc />
-                    </div>
-                </Route>
-
+                <Route path="/home" component={Dashboard}/>
+                <Route exact path="/prof" component={Prof}/>
+                <Route path="/products" component={Products} />
+                <Route path="/test" component={Test} />
+                <Route path="/doc" component={Doc} />
             </Switch>
         </Router>
     )
