+++ src/componentsUtils/Docs/Doc.js
@@ -5,8 +5,8 @@ import CssBaseline from '@material-ui/core/CssBaseline';
 import Grid from '@material-ui/core/Grid';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
-import { makeStyles } from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
+import theme from './theme'
 
 // function Copyright() {
 //   return (
@@ -21,41 +21,10 @@ import Container from '@material-ui/core/Container';
 //   );
 // }
 
-const useStyles = makeStyles((theme) => ({
-  icon: {
-    marginRight: theme.spacing(2),
-  },
-  heroContent: {
-    backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(8, 0, 6),
-  },
-  heroButtons: {
-    marginTop: theme.spacing(4),
-  },
-  cardGrid: {
-    paddingTop: theme.spacing(8,0,4),
-    paddingBottom: theme.spacing(8,0,4),
-  },
-  card: {
-    height: '100%',
-    display: 'flex',
-    flexDirection: 'column',
-  },
-  cardMedia: {
-    paddingTop: '56.25%', // 16:9
-  },
-  cardContent: {
-    flexGrow: 1,
-  },
-  footer: {
-    backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(6),
-  },
-}));
 
 
 export default function Doc() {
-  const classes = useStyles();
+  const classes = theme();
 
   return (
     <React.Fragment>
@@ -111,7 +80,7 @@ export default function Doc() {
             </div>
           </Container>
         </div>
-        
+
       </main>
       {/* Footer */}
       {/* <footer className={classes.footer}>
