+++ src/componentsUtils/Products/Products.js
@@ -9,8 +9,8 @@ import CssBaseline from '@material-ui/core/CssBaseline';
 import Grid from '@material-ui/core/Grid';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
-import { makeStyles } from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
+import theme from './theme.js'
 
 
 // function Copyright() {
@@ -26,45 +26,11 @@ import Container from '@material-ui/core/Container';
 //   );
 // }
 
-const useStyles = makeStyles((theme) => ({
-  icon: {
-    marginRight: theme.spacing(2),
-  },
-  heroContent: {
-    backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(8, 0, 6),
-  },
-  heroButtons: {
-    marginTop: theme.spacing(4),
-  },
-  cardGrid: {
-    paddingTop: theme.spacing(8),
-    paddingBottom: theme.spacing(8),
-  },
-  appBar: {
-    backgroundColor : '#2e7d32',
-  },
-  card: {
-    height: '100%',
-    display: 'flex',
-    flexDirection: 'column',
-  },
-  cardMedia: {
-    paddingTop: '56.25%', // 16:9
-  },
-  cardContent: {
-    flexGrow: 1,
-  },
-  footer: {
-    backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(6),
-  },
-}));
 
 // const cards = [1, 2, 3, 4, 5, 6];
 
 export default function Products() {
-  const classes = useStyles();
+  const classes = theme();
 
   return (
     <React.Fragment>
@@ -84,7 +50,7 @@ export default function Products() {
               Our recomendations
             </Typography>
             <Typography variant="h5" align="center" color="textSecondary" paragraph>
-              Our machine learning models determine what's the right investment strategy for you. Your Risk tolerance level is considered as a key criteria 
+              Our machine learning models determine what's the right investment strategy for you. Your Risk tolerance level is considered as a key criteria
             </Typography>
             <div className={classes.heroButtons}>
               <Grid container spacing={2} justify="center">
@@ -175,7 +141,7 @@ export default function Products() {
                   </CardActions>
                 </Card>
               </Grid>
-              
+
           </Grid>
         </Container>
       </main>
